requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiogram.contrib==1.1.4
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
soupsieve==2.8
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
dotenv
aiohttp
django
djangorestframework